requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
django-filter>=25.1
djangorestframework-simplejwt>=5.3.0
